drivers projects: fix CodeQL-flagged defects
Resolve static-analysis findings reported by CodeQL across drivers and
example projects. All fixes are behavior-preserving except where the
previous behavior was itself a bug (sd.c, adin1110 ethertype print).

- sd-card: sd.c: guard against uint32_t underflow when response_len is
  0 (compare response_len > 1 instead of response_len - 1 > 0), which
  previously passed a huge length to memset.

- Widen narrow loop indices compared against wider bounds so the loop
  cannot wrap: ad5706r, ad9517, ad9523, ad9528, max96792_csi,
  iio_adis, ade9113, adpd410x, stm32 dma/gpdma/spi/xspi, nmea_ubx,
  eval-adxl367z dummy_example.

- Match printf/snprintf format specifiers to their argument types
  (cast uint32_t/int32_t args to (unsigned) long for %l): ad9144,
  ad9523, nmea_ubx, iio_adis, ad413x, ad469x_evb, ad713x_fmcz,
  ad796x_fmcz, ad9656_fmc, aducm3029_flash_demo, cn0561, cn0565,
  pulsar-adc. ad413x dropped a stray argument; adin1110 printed a
  two-byte ethertype array with a single %02X - now prints both bytes.

- Cast an operand to the wider type before multiplying to avoid
  overflow before the widening conversion: iio_dac_demo, ad9545,
  adf4156, adf4157, adv7511 cf_hdmi.

- cn0565 eit: check sscanf return against the expected conversion
  count (== 1); the uint8_t result stored EOF as 255, so parse
  failures were treated as success.

+++ drivers/dac/dac_demo/iio_dac_demo.c
@@ -126,7 +126,7 @@ int dac_submit_samples(struct iio_device_data *dev_data)
 
 		while (get_next_ch_idx(desc->active_ch, ch, &ch)) {
 			uint16_t* ch_buffer = (uint16_t*)(desc->loopback_buffers +
-							  (ch * desc->loopback_buffer_len *
+							  ((size_t)ch * desc->loopback_buffer_len *
 							   sizeof(uint16_t) / sizeof(ch_buffer)));
 			ch_buffer[i] = data[k++];
 		}
@@ -169,7 +169,7 @@ int dac_demo_trigger_handler(struct iio_device_data *dev_data)
 	/* Write data to the device one sample/channel at a time */
 	while (get_next_ch_idx(desc->active_ch, ch, &ch)) {
 		uint16_t* ch_buffer = (uint16_t*)(desc->loopback_buffers +
-						  (ch * desc->loopback_buffer_len *
+						  ((size_t)ch * desc->loopback_buffer_len *
 						   sizeof(uint16_t) / sizeof(ch_buffer)));
 		ch_buffer[i] = data[k++];
 	}
