+++ projects/cn0561/src/cn0561.c
@@ -315,15 +315,15 @@ int main()
 
 	for (i = 0; i < CN0561_FMC_SAMPLE_NO; i++) {
 		j = 0;
-		printf("%lu: ", i);
+		printf("%lu: ", (unsigned long)i);
 		while (j < CN0561_FMC_CH_NO) {
 			adc_buffer[CN0561_FMC_CH_NO * i + j] &= 0xffffff00;
 			adc_buffer[CN0561_FMC_CH_NO * i + j] >>= 8;
 			data = lsb * (int32_t)adc_buffer[CN0561_FMC_CH_NO * i + j];
 			if (data > 4.095)
 				data = data - 8.192;
-			printf("CH%lu: 0x%08lx = %+1.5fV ", j,
-			       adc_buffer[CN0561_FMC_CH_NO * i + j], data);
+			printf("CH%lu: 0x%08lx = %+1.5fV ", (unsigned long)j,
+			       (unsigned long)adc_buffer[CN0561_FMC_CH_NO * i + j], data);
 			if (j == (CN0561_FMC_CH_NO - 1))
 				printf("\n");
 			j++;
